Add CSRF checks.

+++ controller/main_controller.php
@@ -137,6 +137,11 @@ class main_controller
 			return $this->error('AJAX_SHOUTBOX_ERROR', 'AJAX_SHOUTBOX_NO_PERMISSION', 403);
 		}
 
+		if (!check_form_key('ajaxshoutbox_posting', 3600 * 12)) // Allow 12 hours.
+		{
+			return $this->error('AJAX_SHOUTBOX_ERROR', 'FORM_INVALID', 500);
+		}
+
 		if ($this->request->is_ajax())
 		{
 			$message = $msg     = trim($this->request->variable('text_shoutbox', '', true));
@@ -195,6 +200,12 @@ class main_controller
 		{
 			return $this->error('AJAX_SHOUTBOX_ERROR', 'AJAX_SHOUTBOX_MISSING_ID', 500);
 		}
+
+		if (!check_form_key('ajaxshoutbox_delete_' . $id)) // Every delete form has its unique form key, based on ID.
+		{
+			return $this->error('AJAX_SHOUTBOX_ERROR', 'FORM_INVALID', 500);
+		}
+
 		try
 		{
 			$this->delete->delete_post($id);
@@ -330,13 +341,15 @@ class main_controller
 		$username = str_replace('./../../', generate_board_url() . '/', $username); // Fix paths
 		$username = str_replace('./../', generate_board_url() . '/', $username); // Fix paths
 
-		return array(
+		$result = array(
 			'id'      => $row['shout_id'],
 			'user'    => $username,
 			'date'    => $this->user->format_date($row['post_time'], $this->user->data['user_ajaxshoutbox_format']),
 			'message' => $text,
 			'delete'  => ($this->auth->acl_get('m_shoutbox_delete') || ($this->auth->acl_get('u_shoutbox_delete') && $row['user_id'] == $this->user->data['user_id'])),
 		);
+
+		return array_merge($result, $this->add_form_key('ajaxshoutbox_delete_' . $row['shout_id']));
 	}
 
 	/**
@@ -363,5 +376,30 @@ class main_controller
 		));
 
 		return $json;
+	}
+
+	/**
+	 * Add a secret token and returns it as array with creation_time and form_token.
+	 *
+	 * Based on phpBB's add_form_key. Compatible with check_form_key.
+	 *
+	 * IMPORTANT: The original event is not included, because the form is build before the event,
+	 * while this function returns the (Possible modified) data after the event.
+	 *
+	 * @param string $form_name The name of the form; has to match the name used in check_form_key, otherwise no
+	 *                          restrictions apply
+	 *
+	 * @return array
+	 */
+	function add_form_key($form_name)
+	{
+		$now = time();
+		$token_sid = ($this->user->data['user_id'] == ANONYMOUS && !empty($this->config['form_token_sid_guests'])) ? $this->user->session_id : '';
+		$token = sha1($now . $this->user->data['user_form_salt'] . $form_name . $token_sid);
+
+		return array(
+			'creation_time' => $now,
+	        'form_token' => $token,
+		);
 	}
 }
